Restructured application to move UI to its own folder, build still works
